requirements.txt
PyQt6>=6.6.0
python-mpv>=1.0.6
natsort>=8.4.0
